--- src/pages/mytrip/ViewAllPage/style.ts
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-export const CardList = styled.ol`
-  margin-top: 24px;
-  padding: 0 10px 10px;
-
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-`;
